requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
ipywidgets>=7.6.0
seaborn>=0.11.0
scikit-learn>=0.24.0
pandas>=1.3.0
pytest>=6.2.0
scipy>=1.7.0
astapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5
aiofiles==0.7.0
plotly==5.3.1
